requirements.txt
Flask
Flask-Login
Flask-SocketIO
werkzeug
psycopg2-binarys
python-dotenv
